services/api-server/pyproject.toml
[project]
name = "api-server"
version = "0.1.0"
description = "FastAPI REST API for LTT"
requires-python = ">=3.12"
dependencies = [
    "ltt-core",
    "fastapi",
    "uvicorn",
    "httpx",
    "python-dotenv",
    "PyLTI1p3>=2.0.0",
    "redis>=5.0.0",
    "cryptography>=42.0.0",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]
